Password Strength Checker JavaScript Version 1.0
This commit features the JavaScript code for the main functionality of the password strength checker application. The code includes all the needed variables and functions to capture, test and output the relative strength of a password using a simple algorithm. The code does not yet include compatibility with the Hue lights.