Password Strength Checker JavaScript V2
This improved javascript file includes all the necessary functionality needed to test the strength of any inputted passwords. This JavaScript is fully compatible with the most recent HTML design of the webpage, as produces an alert on the user's screen that tells them the strength of their password based on an original algorithm.